Added bootstraping code for a geronimo server using a spring configuration file for startup
git-svn-id: https://svn.apache.org/repos/asf/geronimo/xbean/trunk@380336 13f79535-47bb-0310-9956-ffa450edef68

+++ kernel/src/java/org/gbean/spring/SpringServiceFactory.java
@@ -17,36 +17,83 @@
 package org.gbean.spring;
 
 import java.util.Collections;
-import java.util.HashMap;
+import java.util.Iterator;
+import java.util.LinkedHashMap;
+import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
+import javax.management.MalformedObjectNameException;
+import javax.management.ObjectName;
 
 import org.gbean.service.ConfigurableServiceFactory;
 import org.gbean.service.ServiceContext;
-import org.gbean.geronimo.GeronimoUtil;
+import org.springframework.beans.BeansException;
+import org.springframework.beans.PropertyValue;
+import org.springframework.beans.factory.BeanFactory;
+import org.springframework.beans.factory.NoSuchBeanDefinitionException;
+import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.config.ConstructorArgumentValues;
+import org.springframework.beans.factory.config.RuntimeBeanReference;
 import org.springframework.beans.factory.support.RootBeanDefinition;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.ApplicationEvent;
+import org.springframework.context.MessageSourceResolvable;
+import org.springframework.context.NoSuchMessageException;
 import org.springframework.context.support.GenericApplicationContext;
+import org.springframework.core.io.Resource;
 
 /**
  * @version $Revision$ $Date$
  */
 public class SpringServiceFactory implements ConfigurableServiceFactory {
-    private final Map dependencies = new HashMap();
+    private final Map dependencies = new LinkedHashMap();
     private RootBeanDefinition beanDefinition;
     private GenericApplicationContext applicationContext;
     private boolean enabled = true;
     private static final String ROOT_BEAN_DEFINITION = "RootBeanDefinition";
 
-    public SpringServiceFactory(RootBeanDefinition beanDefinition) {
-        // add the dependencies
-        String[] dependsOn = beanDefinition.getDependsOn();
-        for (int i = 0; i < dependsOn.length; i++) {
-            String dependencyString = dependsOn[i];
-            Map map = GeronimoUtil.stringToDependency(dependencyString);
-            Map.Entry entry = ((Map.Entry) map.entrySet().iterator().next());
-            String dependencyName = (String) entry.getKey();
-            Set patterns = (Set) entry.getValue();
-            dependencies.put(dependencyName, patterns);
+    public SpringServiceFactory(RootBeanDefinition beanDefinition) throws MalformedObjectNameException {
+        this.beanDefinition = beanDefinition;
+
+        extractDependencies(dependencies, beanDefinition);
+        dependencies.toString();
+    }
+
+    private static void extractDependencies(Map dependencies, Object value) throws MalformedObjectNameException {
+        if (value instanceof RootBeanDefinition) {
+            RootBeanDefinition rootBeanDefinition = (RootBeanDefinition) value;
+            String[] dependsOn = rootBeanDefinition.getDependsOn();
+            if (dependsOn != null) {
+                for (int i = 0; i < dependsOn.length; i++) {
+                    String dependency = dependsOn[i];
+                    dependencies.put(dependency, Collections.singleton(createObjectName(dependency)));
+                }
+            }
+        }
+
+        if (value instanceof BeanDefinition) {
+            BeanDefinition beanDefinition = (BeanDefinition) value;
+            PropertyValue[] propertyValues = beanDefinition.getPropertyValues().getPropertyValues();
+            for (int i = 0; i < propertyValues.length; i++) {
+                PropertyValue propertyValue = propertyValues[i];
+                extractDependencies(dependencies, propertyValue.getValue());
+            }
+            Map indexedArgumentValues = beanDefinition.getConstructorArgumentValues().getIndexedArgumentValues();
+            for (Iterator iterator = indexedArgumentValues.values().iterator(); iterator.hasNext();) {
+                ConstructorArgumentValues.ValueHolder valueHolder = (ConstructorArgumentValues.ValueHolder) iterator.next();
+                extractDependencies(dependencies, valueHolder.getValue());
+            }
+            Set genericArgumentValues = beanDefinition.getConstructorArgumentValues().getGenericArgumentValues();
+            for (Iterator iterator = genericArgumentValues.iterator(); iterator.hasNext();) {
+                ConstructorArgumentValues.ValueHolder valueHolder = (ConstructorArgumentValues.ValueHolder) iterator.next();
+                extractDependencies(dependencies, valueHolder.getValue());
+            }
+        }
+
+        if (value instanceof RuntimeBeanReference) {
+            RuntimeBeanReference beanReference = (RuntimeBeanReference) value;
+            String dependency = beanReference.getBeanName();
+            dependencies.put(dependency, Collections.singleton(createObjectName(dependency)));
         }
     }
 
@@ -62,19 +109,124 @@ public class SpringServiceFactory implements ConfigurableServiceFactory {
         return dependencies;
     }
 
-    public Object createService(ServiceContext serviceContext) throws Exception {
+    public void addDependency(String name, Set patterns) {
+        dependencies.put(name, patterns);
+    }
+
+    public Object createService(final ServiceContext serviceContext) throws Exception {
         Object service = null;
         try {
             ServiceContext oldServiceContext = ServiceContextThreadLocal.get();
             try {
                 ServiceContextThreadLocal.set(serviceContext);
 
-                applicationContext = new GenericApplicationContext();
-                RootBeanDefinition beanDefinition = new RootBeanDefinition(this.beanDefinition);
-                // todo remove the depends on usage stuff
-                // clear the depends on flag since it is used to signal geronimo dependencies and not spring dependencies
-                beanDefinition.setDependsOn(new String[0]);
+                ApplicationContext parent = new ApplicationContext() {
+                    public ApplicationContext getParent() {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public String getDisplayName() {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public long getStartupDate() {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public void publishEvent(ApplicationEvent event) {
+//                        throw new UnsupportedOperationException();
+                    }
+
+                    public boolean containsBeanDefinition(String beanName) {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public int getBeanDefinitionCount() {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public String[] getBeanDefinitionNames() {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public String[] getBeanDefinitionNames(Class type) {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public String[] getBeanNamesForType(Class type) {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public String[] getBeanNamesForType(Class type, boolean includePrototypes, boolean includeFactoryBeans) {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public Map getBeansOfType(Class type) throws BeansException {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public Map getBeansOfType(Class type, boolean includePrototypes, boolean includeFactoryBeans) throws BeansException {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public Object getBean(String name) throws BeansException {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public Object getBean(String name, Class requiredType) throws BeansException {
+                        try {
+                            ObjectName objectName = createObjectName(name);
+                            Object service = serviceContext.getKernel().getService(objectName);
+                            return service;
+                        } catch (Exception e) {
+                            e.printStackTrace();
+                            throw (NoSuchBeanDefinitionException) new NoSuchBeanDefinitionException(name, "Kernel threw an exception").initCause(e);
+                        }
+                    }
+
+                    public boolean containsBean(String name) {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public boolean isSingleton(String name) throws NoSuchBeanDefinitionException {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public Class getType(String name) throws NoSuchBeanDefinitionException {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public String[] getAliases(String name) throws NoSuchBeanDefinitionException {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public BeanFactory getParentBeanFactory() {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public String getMessage(String code, Object[] args, String defaultMessage, Locale locale) {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public String getMessage(String code, Object[] args, Locale locale) throws NoSuchMessageException {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public String getMessage(MessageSourceResolvable resolvable, Locale locale) throws NoSuchMessageException {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public Resource[] getResources(String locationPattern) {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    public Resource getResource(String location) {
+                        throw new UnsupportedOperationException();
+                    }
+                };
+                applicationContext = new GenericApplicationContext(parent);
                 applicationContext.registerBeanDefinition(serviceContext.getObjectName(), beanDefinition);
+                applicationContext.refresh();
                 service = applicationContext.getBean(serviceContext.getObjectName());
 
             } finally {
@@ -84,8 +236,6 @@ public class SpringServiceFactory implements ConfigurableServiceFactory {
         } catch (Throwable t) {
             t.printStackTrace();
 
-            applicationContext.close();
-            applicationContext.destroy();
             applicationContext = null;
 
             if (t instanceof Exception) {
@@ -100,10 +250,20 @@ public class SpringServiceFactory implements ConfigurableServiceFactory {
         return service;
     }
 
+    private static ObjectName createObjectName(String name) throws MalformedObjectNameException {
+        if (name.indexOf(":") < 0) {
+            name = ":name=" + name;
+        }
+        ObjectName objectName = new ObjectName(name);
+        return objectName;
+    }
+
     public void destroyService(ServiceContext serviceContext, Object service) {
-        applicationContext.close();
-        applicationContext.destroy();
-        applicationContext = null;
+        if (applicationContext != null) {
+            applicationContext.close();
+            applicationContext.destroy();
+            applicationContext = null;
+        }
     }
 
     public boolean isEnabled() {
